requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
ollama==0.4.8
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg==3.2.9
pydantic==2.11.4
pydantic_core==2.33.2
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
wheel==0.45.1
